+++ resources/js/Layouts/AuthLayout.vue
